added vundle installation

+++ install.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
 
+if [ ! -d ~/.vim/bundle/Vundle.vim ]; then
+    echo "installing Vundle for plugins"
+    git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+fi
+
 DIR=$(cd "$(dirname "$0")"; pwd)
 
-echo "creating symbolic link ~/.gitconfig from $DIR/git/.config"
-ln -s $DIR/git/.gitconfig ~/.gitconfig
+if [ ! -L ~/.gitconfig ]; then
+    echo "creating symbolic link ~/.gitconfig from $DIR/git/.config"
+    ln -s $DIR/git/.gitconfig ~/.gitconfig
+fi
 
-echo "creating symbolic link ~/.vimrc from $DIR/vim/.vimrc"
-ln -s $DIR/vim/.vimrc ~/.vimrc
+if [ ! -L ~/.vimrc ]; then
+    echo "creating symbolic link ~/.vimrc from $DIR/vim/.vimrc"
+    ln -s $DIR/vim/.vimrc ~/.vimrc
+fi
